rename loggers/wandb_logging to loggers/wandb
