+++ utils/loggers/__init__.py
@@ -7,7 +7,7 @@ from torch.utils.tensorboard import SummaryWriter
 
 from utils.general import colorstr
 from utils.torch_utils import de_parallel
-from utils.loggers.wandb_logging.wandb_utils import WandbLogger
+from utils.loggers.wandb.wandb_utils import WandbLogger
 
 LOGGERS = ('txt', 'tb', 'wandb')  # text-file, TensorBoard, Weights & Biases
